Allow specifying versions for plugin_gems
Previously, it was possible only to install the latest version of a gem.
This allows specifying gem versions with the same array arguments expected
by Scout.install_gem.

For example, the following node configuration will install the latest
version of the docker-api gem:

    default[:scout][:plugin_gems] = ['docker-api']

This configuration, on the other hand, will install version 1.17.0

    default[:scout][:plugin_gems] = [%w(docker-api --version 1.17.0)]