Allow string values in plugin_properties hash
When installing scoutd with Chef under AWS's Opsworks, data bag support is
limited and parameters that we'd like to put in the plugins.properties file need
to be passed via Opsworks's "custom JSON" hashes, which end up in the chef node.
This patch allows plugins.properties to be initialized either via data bags (the
current behavior) or via raw strings.